Update distribution design models directory structure. The schema Distro/$RELEASE/Component makes rendition difficult to control through FLAG_FILTER. There are two nested loops (one to load pre-rendition configuration scripts and other to load design models related to the pre-rendition previously loaded) and we need to deside which one of them does use the FLAG_FILTER.
    - If we use FLAG_FILTER to customize the first loop (the one
      loading pre-rendition configuration scripts) the best we can
      customize is which component is rendered but not which file
      inside the component because that is controlled by the second
      loop.

    - If we use FLAG_FILTER  to customize the second loop (the one
      loading design models, related to previously loaded
      pre-rendition configuration scrtip), the best we can do is
      customize which file names inside all components will be
      rendered, but not which file in which component.

    - If we use FLAG_FILTER in obth first and second loop makes no
      much useful thing. In this case, we are trying to apply a common
      search pattern to two different path informations which probably
      get in conflict one another since the regular expression pattern
      point of view.

So, go back to the previous configuration where release and
architecture information are set at the end of the path (e.g.,
Distro/Component/$RELEASE).

Another point of this configuration is that release-specific
information can be supplied by means of `--release' option. There is
no need to have one release-specific directory inside design models to
produce design models on different artistic motifs. Nevertheless it
could be required if file names, of files produced, change from one
major release to another; or maybe not, since a tune up script could
be set to fix the name issue or something similar.