Remove Trunk, Branches and Tags chapters from Tcar-fs.
Previous to this commit, the texinfo documentation models for Tcar-fs
documentation manual were organized using Trunk, Branches and Tags as
chapters and all directories inside it as sections. This was fine for
a conventional trunk, branches, tags layout. But it isn't for a Git
layout where such structure isn't required.

In this commit, I remove the Trunk, Branches and Tags chapters and
start using Identity/ and Scripts/ as chapters instead. This to
reflect the same structure provided by Git. The Locales/ and
Documentation/ directories hadn't information at time of this commit
so I didn't added now.