Remove trunk/ from path references inside scripts.
Previous to this commit, references to the repository directory
structure were taking place by combining the TCAR_WORKDIR environment
variable and a path starting with `trunk/' (e.g.,
${TCAR_WORKDIR}/trunk/Identity/Models/Brands). This was fine until we
decided to start using Git instead of Subversion as main version
control system inside the repository. Git handles branches and tags in
a way completely different than Subversion does, so it is necessary to
change the way we make references to repository directory structure
from automation scripts. This in order to support both Git and
Subversion layouts.

In this commit, the `trunk/' string is no longer prefixed on paths but
suffixed before the TCAR_WORKDIR environment variable. To know whether
to suffix the `trunk/' string or not the first time the centos-art
prepare script is executed, the prepare functionality verifies the
existence of a `trunk' directory in the first level of the working
copy and adds the string based on such a fact. If the `trunk'
directory doesn't exist, the `trunk/' string won't be suffixed to
TCAR_WORKDIR environment variable.

Changing repository directory layout this way affects the following:

 - Static includes used inside Texinfo documentation to load common
   contents (e.g., licenses, etc.). With this change, documentation
   manuals won't be rendered if the first directory level of the
   repository is a trunk-branches-tags layout. This is because the
   static paths used as argument to `@include' commands inside Texinfo
   documentation source files were changed to work with Git layout,
   where there isn't any trunk/, branches/ and tags/ directories as
   first level in the repository directory structure.

 - The current layout of The CentOS Artowork Repository File System
   (tcar-fs) texinfo documentation manual. There is no longer Trunk,
   Tags or Branches chapters in Texinfo documentation but Identity,
   Documentation, Locales, and Scripts chapters, instead.

There are other minor changes that are also committed here and I want
to mention briefly:

 - Update code description set as comments inside shell scripts.

 - Update the `To know more ...' message. Previous to this commit, to
   find help about one of the functionalities, the command output in a
   `To know more ...' message used the help command and the
   functionality path. This commit implements one `--help' option for
   some of the specific functionalities available. This option calls
   the help functionality internally.

 - The =REPO_TLDIR=, =REPO_HOME= and =TCAR_WORKDIR= translation
   markers, all point to the value of TCAR_WORKDIR environment
   variable.

- The svn functionality doesn't applies anymore and should be renamed
  to something more appropriate (e.g., vcs) and subdivided in svn and
  git functionalities so it be possible to manage both subversion and
  git repository versioning tasks, respectively.

- The locale functionality becomes more specific when uses the
  cli_getFilesList function to build the list of directories that need
  to be localized when the `--update' option is provided. It no longer
  uses the regular expression '/[^.]${FLAG_FILTER}' as pattern but
  '${ACTIONVAL}/${FLAG_FILTER}' which is more specific.

- Include `set spell' to Vim's default configuration file in order to
  activate the Vim's spelling checking functionality to all files we
  edit.

- Rename function name from cli_runFunEnvironment to
  cli_runFnEnvironment on all sh files.